FractalTools.jl

FractalTools.IFS — Type

struct IFS{T1, T2}

Iterated fucntion sytem (IFS)

Fields

ws::Any

probs::Any

FractalTools.Interpolant — Type

struct Interpolant{T1<:IFS, T2<:(AbstractArray{var"#s12",1} where var"#s12"<:Interval), T3} <: AbstractInterpolant

One dimensional fractal interpolant

Fields

ifs::IFS

intervals::AbstractArray{var"#s12",1} where var"#s12"<:Interval

itp::Any

FractalTools.Interval — Type

struct Interval{T<:Real}

Real closed interval

Fields

lb::Real

ub::Real

FractalTools.Transformation — Type

struct Transformation{T1, T2}

Affine transformation

Fields

A::Any

b::Any

Base.:∈ — Method

∈(x, I)

Returns true if x is in I.

Example

julia> I = Interval(1, 2)
Interval{Int64}(1, 2)

julia> 1.5 ∈ I 
true

FractalTools.fif — Function

fif(x, y, d)
fif(x, y, d, f0)
fif(x, y, d, f0, niter)

Interpolates the data pairs (xi, yi) for xi ∈ x and yi ∈ y. f0 is the initial function and niter is the number of iterations.

FractalTools.gethiddenifs — Method

gethiddenifs(x, y, z, d, h, l, m)

Returns the IFS corresponding to interpolation data x, y. d is the scaling parameters. See the notes on docs

FractalTools.getifs — Method

getifs(x, y, d)

Returns the IFS corresponding to interpolation data x, y. d is the scaling parameters.

FractalTools.getintervals — Method

getintervals(x)

Returns the just-touching intervals

FractalTools.hiddenfif — Method

hiddenfif(x, y, z, d, h, l, m; f0, niter)

Interpolates the data pairs (xi, yi) for xi ∈ x and yi ∈ y. f0 is the initial function and niter is the number of iterations.

FractalTools.wrap — Method

wrap()

Functional that takes initiali function func and returns a wrapped function.

FractalTools.wraphidden — Method

wraphidden()

Functional that takes initiali function func and returns a wrapped function.