FractalTools.jl

FractalTools.Attractor — Type

struct Attractor{T, S, R}

Attractor of IFS type

Fields

ifs::Any

IFS of Attractor

alg::Any

Type of algorithm to be used to compute attractor(Options are DetAlg and RandAlg

initset::Any

Initial set of attractor

set::Any

Set of the attractor

numiter::Int64

Number of iterations

parallel::Bool

Sequential or parallel

FractalTools.DetAlg — Type

struct DetAlg

A type signifying that deterministic algorithm is used when calculating the attractor of and IFS.

FractalTools.IFS — Type

struct IFS{T1, T2}

Iterated fucntion sytem (IFS)

Fields

ws::Any

probs::Any

FractalTools.Interpolant — Type

struct Interpolant{T1<:IFS, T2<:(AbstractArray{var"#s37",1} where var"#s37"<:Interval), T3} <: AbstractInterpolant

One dimensional fractal interpolant

Fields

ifs::IFS

intervals::AbstractArray{var"#s37",1} where var"#s37"<:Interval

itp::Any

FractalTools.Interval — Type

struct Interval{T<:Real}

Real closed interval

Fields

lb::Real

ub::Real

FractalTools.RandAlg — Type

struct RandAlg

A type signifying that random algorithm is used when calculating the attractor of and IFS.

FractalTools.Transformation — Type

struct Transformation{T1, T2}

Affine transformation

Fields

A::Any

b::Any

Base.:∈ — Method

∈(x, I)

Returns true if x is in I.

Example

julia> I = Interval(1, 2)
Interval{Int64}(1, 2)

julia> 1.5 ∈ Interval(1, 2)
true

FractalTools.Fern — Method

Fern()

Constructs and IFS for a fern.

FractalTools.Sierpinski — Method

Sierpinski()

Conctructs an IFS for Sierpinski triangle.

FractalTools.Square — Method

Square()

Constructs and IFS for a sqaure.

FractalTools.Tree — Method

Tree()

Constructs and IFS for a fractal tree.

FractalTools.attractor — Method

attractor(ifs, initset; alg, kwargs...)

Computes the attractor of ifs. If alg is of type DetAlg, the deterministic algorithm is used. If alg is of type RandAlg, random algorithm is used. kwargs may include

numiter::Int : Number of iterations to used to calcuate the attractor (defaults to 10)

numtransient::Int : Number of transient iterations to used to calcuate a transient set. When the transient set is constructed, the computation of attractor is continued with distributed computation if alg is RandAlg and parallel is true. (defaults to 10)

parallel::Bool: If true, the attractor is computed using distrbuted computation. (defaults to false)

placedependent::Bool : If true, place dependent attractor is computed if α and β is given accordingly. (default to false)

α::AbstractVector : Place-dependent probility coefficient(defaults to nothing)

β::AbstractVector : Place-dependent probility coefficient. (default to nothing)

FractalTools.boundarypoints — Method

boundarypoints(p1, p2, p3; numpoints)

Returns boundary points of on the edges of the triangle p1, p2, p3. numpoints is the number of points on the edges of the triangle.

FractalTools.box — Method

box(p1, p2, p3)

Returns the matrix A and a vector b such that for any point p the transformation T(p) = A * p + b moves p into the bounding box of the triangule formed by the points p1, p2, p3.

FractalTools.coefficients — Method

coefficients(x, y, d)

Returns the coefficients of the transformations of the ifs of the fractal interpolation.

FractalTools.coefficients — Method

coefficients(mesh, z, α, region_type)

Computes coefficients of the iterated function system given the two dimensional interpolation domain mesh and corresponding z with the vertical scaling factor α.

FractalTools.coefficients — Method

coefficients(mesh, z, α, region_type)

Computes coefficients of the iterated function system given the two dimensional interpolation domain mesh and corresponding z with the vertical scaling factor α.

FractalTools.coefficients2equationsystem — Method

coefficients2equationsystem(coeffs, α)

Returns the matrix, vector corresponding to the IFS given by coeffs and the vertical scaling factors α.

FractalTools.coefficients2transforms — Method

coefficients2transforms(coeffs, α)

Converts to equation system given with coeffs and vertical scaling factors α7 into transformations of the IFS.

FractalTools.coefficients2transforms — Method

coefficients2transforms(coeffs, α)

Converts to equation system given with coeffs and vertical scaling factors α7 into transformations of the IFS.

FractalTools.contfactor — Method

contfactor(ifs::IFS)

Returns the contraction factor of IFS.

FractalTools.contfactor — Method

contfactor(w)

Returns contraction factor of w. Contraction factor is computed as the norm of w.A.

FractalTools.contraction_factors — Method

contraction_factors(coeffs, α)

Computes contraction factors of the affine transformations of the ifs given with coeffs and the verticaling factors α. Contraction factors are calculated to be the eigenvalues of the coeffs.

FractalTools.detalg — Method

detalg(ifs, initset; numiter, parallel)

Computes the attractor of ifs with deterministic algorithm.numiter is number of iterations. (Defaults to 10). If parallel is true, attractor is computed via parallel computation.

FractalTools.dimension — Method

dimension(ifs::IFS)

Returns dimension of ifs.

FractalTools.dimension — Method

dimension(w)

Returns dimension of w.

FractalTools.disperse — Method

disperse(trig, npoints)

Returns a vector of random points that are dispersed into trig. npoints is the number of points to be dispersed.

FractalTools.fif — Method

fif(x, y; d, f0, niter)

Interpolates the data pairs (xi, yi) for xi ∈ x and yi ∈ y. f0 is the initial function and niter is the number of iterations.

FractalTools.find_apex — Method

find_apex(mesh, coordinates=false)

Finds the indices of the apex points of the mesh. If coordinates is true, the coordinates of the apex points are returned.

FractalTools.find_region — Method

find_region(regions, px, py)

Finds triangular region in regions to which the point px belongs.

FractalTools.findouttriangle — Method

findouttriangle(pnts3d)

Returns the bounding triangle of the points pnts3d.

FractalTools.fis — Method

fis(mesh, z, α=0.1, func0=(xi,yi)->0., num_iter=15, get_coeffs=false)

Fractal interpolaton surface for the interpolaton domain given with mesh and associated z with the vertical scling factors α, initial function func0, number of iterations num_iter. If get_coeffs is true, coefficients of the IFS orresponding to dataset mesh and z are returned.

FractalTools.fis — Method

fis(mesh, z, α=0.025, func0=(xi,yi)->0., num_iter=15, get_coeffs=false)

Fractal interpolaton surface for the interpolaton domain given with mesh and associated z with the vertical scling factors α, initial function func0, number of iterations num_iter. If get_coeffs is true, coefficients of the IFS corresponding to dataset mesh and z are returned.

FractalTools.getdata — Method

getdata(f, trig, npts)

Returns a three-dimensional interpolation data pnts. pnts is a vector of three-dimensional points pi = Point(xi, yi, zi) where xi and yi are from the dispersed points and zi = f(xi, yi).

FractalTools.gethiddenifs — Method

gethiddenifs(x, y, z, d, h, l, m)

Returns the IFS corresponding to interpolation data x, y. d is the scaling parameters. See the notes on docs

FractalTools.getifs — Method

getifs(x, y, d)

Returns the IFS corresponding to interpolation data x, y. d is the scaling parameters.

FractalTools.getintervals — Method

getintervals(x)

Returns the just-touching intervals

FractalTools.getpoint — Method

getpoint(tri; maxiter)

Returns a random valid point in tri. maxiters is the number of iteration while finding the point.

FractalTools.gettransform — Function

gettransform(outtrig, intrig)
gettransform(outtrig, intrig, α)

Returns a named tuple of A andbsuch thatL(x) = A * x + bmapsouttrigtointrig`.

FractalTools.hiddenfif — Method

hiddenfif(x, y, z; d, h, l, m, f0, niter)

Interpolates the data pairs (xi, yi) for xi ∈ x and yi ∈ y. f0 is the initial function and niter is the number of iterations.

FractalTools.hiddenfis — Method

hiddenfis(mesh, z, α=0.1, func0=(xi,yi)->0., num_iter=15, get_coeffs=false)

Fractal interpolaton surface for the interpolaton domain given with mesh and associated z with the vertical scling factors α, initial function func0, number of iterations num_iter. If get_coeffs is true, coefficients of the IFS orresponding to dataset mesh and z are returned.

FractalTools.hiddenfis — Method

hiddenfis(mesh, z, α=0.025, func0=(xi,yi)->0., num_iter=15, get_coeffs=false)

Fractal interpolaton surface for the interpolaton domain given with mesh and associated z with the vertical scling factors α, initial function func0, number of iterations num_iter. If get_coeffs is true, coefficients of the IFS corresponding to dataset mesh and z are returned.

FractalTools.integrate — Method

integrate(x, y, z; d, h, l, m)

I = ∫f(x)dx

Numerically integrates the interpolation function for interpolation points (x, y). x is a domain point, y is the range point. d is the vertical scaling factor (0., 1.)

FractalTools.integrate — Method

integrate(x::AbstractVector{T}, y::AbstractVector{T}, g::Function;
    d=0.001) where {T<:Real}

Computes the numerical integration $I = ∫g(u)f(u)du$ where f is the interpolation function for the points (x, y). g is any single variable function. d is the vertical scaling factor.

FractalTools.integrate — Method

integrate(x, y; d)

I = ∫f(x)dx

Numerically integrates the interpolation function for interpolation points (x, y). x is a domain point, y is the range point. d is the vertical scaling factor (0., 1.)

FractalTools.integrate — Method

integrate(mesh::PyCall.PyObject, z::AbstractVector; α::Real=0.001, region_type::String="triangular")

I = ∫f(x,y)dxdy

Numerically integrates interpolation function for interpolation points (mesh, zi) where mesh is domain, and zi is the corresponding point in the range. α is vertical scaling factor between (0., 1.). region_type determines the geometry of the regions. triangular and polygonal are the available geometrical regions.

FractalTools.interpolate — Method

interpolate(pnts3d; α, f0, maxiters, gettransforms)

Returns a fractal surface interpolation function that interpolates pnts3d.

FractalTools.isvalidpoint — Method

"

isvalidpoint(pnt, trig)

Returns true if pnt is in trig.

FractalTools.linepoint — Method

linepoint(p1, p2; numpoints)

Returns the points of the edge whose end points are p1 and p2. numpoints is the number of points.

FractalTools.proj_onto_line — Function

proj_onto_line(px, p1, p2, minimizer::Bool=true)

Projects the point px onto the line passing the points p1 ans p2. If back_project is true, minimizer of the disatance of the point to the line is returned instead of the projection point.

FractalTools.proj_onto_triangle — Function

proj_onto_triangle(px, triangle, all)

Projects to the poits px on the nearest boundary of the triangle triangle. If all is true, then the projections of the point px is all three edges of the triangle is returned.

FractalTools.project — Method

project(pnts3d)

Returns the two-dimensional projections of the three-dimensional points pnts3d

FractalTools.project — Method

project(msh3)

Returns a two dimensional mesh whose points are the projections of msh3.

FractalTools.randalg — Method

randalg(ifs, initset; numiter, numtransient, parallel, placedependent, α, β)

Computes the attractor of ifs with random algorithm.numiter is number of iterations. (Defaults to 100). numtransient is the number of transient iterations. If parallel is true, attractor is computed via parallel computation. If placedependent is true, the probabilties of the ifs are dependent on the coordinates x. This dependency p(x) is given via the parameters α and β where p(x) = α x + β.

FractalTools.triangulate — Method

triangulate(pnts3d)

Returns a tuple of a Delaunay tessellation and a correspoding three dimensional mesh.

FractalTools.wrap — Method

wrap()

Functional that takes initiali function func and returns a wrapped function.

FractalTools.wraphidden — Method

wraphidden()

Functional that takes initiali function func and returns a wrapped function.